* revamp multi-tile building handling, allow 2x2 1x3 etc sized buildings

+++ src/client/campaign/cp_base.h
@@ -212,9 +212,9 @@ void B_SetCurrentSelectedBase(const base_t *base);
 qboolean B_AssembleMap(const base_t *base);
 
 /* building functions */
-#define B_IsTileBlocked(base, x, y) (base)->map[(y)][(x)].blocked
+#define B_IsTileBlocked(base, x, y) (base)->map[(int)(y)][(int)(x)].blocked
 #define B_IsBuildingBuiltUp(building) (!(building)->timeStart || (building)->timeStart + (building)->buildTime <= ccs.date.day)
-#define B_GetBuildingAt(base, x, y) (base)->map[(y)][(x)].building
+#define B_GetBuildingAt(base, x, y) (base)->map[(int)(y)][(int)(x)].building
 
 buildingType_t B_GetBuildingTypeByCapacity(baseCapacities_t cap);
